backend/requirements.txt
# G-Mini Agent — Backend Dependencies (Phases 1-3)

# ═══ Phase 1: Core + LLM Chat ═══
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
pyyaml>=6.0
aiofiles>=23.0.0
aiohttp>=3.9.0
discord.py>=2.4.0

# LLM Providers — SDKs actualizados para modelos actuales
openai>=1.82.0            # GPT-5.4 series, Responses API, Realtime
anthropic>=0.52.0         # Claude Opus 4.6, Sonnet 4.6, extended thinking
google-genai>=1.14.0      # Gemini 3.1 Pro, 3 Flash, google-genai SDK

# Token counting
tiktoken>=0.7.0

# Security
keyring>=25.0.0

# Database
aiosqlite>=0.20.0
croniter>=2.0.0

# Logging
loguru>=0.7.0

# Utils
numpy>=1.26.0
Pillow>=10.2.0

# ═══ Phase 2: Vision + Automation ═══
# Screen capture
mss>=9.0.0

# OCR (instalar uno o más)
pytesseract>=0.3.10
easyocr>=1.7.0
# paddleocr>=2.7.0   # Descomentar si se desea PaddleOCR (pesado)

# PC Automation
pyautogui>=0.9.54
pynput>=1.7.6
pyperclip>=1.8.2
browser-use>=0.12.0        # Automatización browser (reemplaza playwright directo)
# playwright se instala como dependencia de browser-use (no usar directo)

# Android (ADB)
pure-python-adb>=0.3.0.dev0

# UI Detection (opcional — requiere modelo YOLO)
ultralytics>=8.3.0
huggingface-hub>=0.34.0
transformers>=4.46.0,<5.0.0
safetensors>=0.4.5
timm>=1.0.11
einops>=0.8.0

# ═══ Phase 3: Voice + Character ═══
# TTS
# melo-tts>=0.1.0    # Descomentar cuando esté disponible en PyPI
elevenlabs>=1.0.0

# STT
faster-whisper>=1.0.0

# Real-Time Voice
websockets>=12.0

# Audio I/O
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad>=2.0.10

# ═══ Phase 6: Gateway (Slack) ═══
slack-sdk>=3.27.0

# ═══ Phase 11: Knowledge Graph ═══
networkx>=3.2
